:hammer: 엔티티 개선 서브 카테고리 엔티티 개선 작업 관련 테스트 코드 작성

+++ src/main/java/com/giyong/community/controller/SubCategoryController.java
@@ -45,7 +45,7 @@ public class SubCategoryController {
     }
 
     @GetMapping("/categorys")
-    public String categories(Integer subCategoryId, Integer page, Pageable pageable, Model m) {
+    public String categories(Long subCategoryId, Integer page, Pageable pageable, Model m) {
         SubCategory subCategory = subCategoryService.findById(subCategoryId);
         Page<MainCategory> mainCategories = mainCategoryService.findAll(pageable);
         m.addAttribute("subCategory", subCategory);
