+++ src/main/java/com/giyong/community/repository/SubCategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SubCategoryRepository extends JpaRepository<SubCategory, Integer> {
+public interface SubCategoryRepository extends JpaRepository<SubCategory, Long> {
     Optional<SubCategory> findBySubCategoryName(String subCategoryName);
 }
